refactor(tests): mirror main service tests into infrastructure domains
