+++ tests/unit/platform/updates/update.service.test.ts
@@ -6,7 +6,7 @@
 import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest';
 import { UpdateService } from '@platform/updates';
 import { UpdateState } from '@platform/config';
-import { createEventBus, createLoggerFactory, createUpdateConfigMock, createWindowServiceMock } from '../../factories/index.js';
+import { createEventBus, createLoggerFactory, createUpdateConfigMock, createWindowServiceMock } from '../../../factories/index.js';
 
 vi.mock('electron', () => ({
   app: {
